Add files via upload
Added function to extract labels and convert them to a string from the AWS response.  Cleaned up the Amazon code as I thought it looked messy.  Commented out a few things that won't be needed for the final project.  Added counter to stop model after 10 images are analyzed.  Found out I was being accrued charges so that was necessary code.